Visual changes + use JQuery UI + drag/drop works
Added some CSS to make the tables have blue headers to match the rest of the page, instead of using the default html ones. Fixed some spacing things. Added JQuery UI as a script for the teacher page in order to get drag-and-drop to function. Added a temporary table for some arbitrary group to show that dragging and dropping works from the ungrouped students table to the group table.